Added Custom authorization policy and user context service

+++ RestaurantApi/Services/RestaurantService.cs
@@ -1,11 +1,15 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using RestaurantApi.Authorization;
 using RestaurantApi.Dtos;
 using RestaurantApi.Dtos.Create;
 using RestaurantApi.Dtos.Update;
 using RestaurantApi.Entities;
 using RestaurantApi.Exceptions;
+using RestaurantApi.Models.Enums;
 using RestaurantApi.Services.Interfaces;
+using System.Security.Claims;
 
 namespace RestaurantApi.Services
 {
@@ -14,15 +18,20 @@ namespace RestaurantApi.Services
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantService> _logger;
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        private readonly IAuthorizationService _authorizationService;
+        private readonly IUserContextService _userContextService;
+        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger, IAuthorizationService authorizationService, IUserContextService userContextService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
+            _userContextService = userContextService;
         }
         public Restaurant Create(CreateRestaurantDto dto)
         {
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = _userContextService.GetUserId;
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
 
@@ -60,7 +69,14 @@ namespace RestaurantApi.Services
             var restaurant = _dbContext.Restaurants
            .FirstOrDefault(r => r.Id == id);
 
-            if (restaurant is null) throw new NotFoundException("Restaurant not found"); ;
+            if (restaurant is null) throw new NotFoundException("Restaurant not found");
+
+            var authorizationResult = _authorizationService.AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             restaurant.Name = dto.Name;
             restaurant.Description = dto.Description;
@@ -77,7 +93,14 @@ namespace RestaurantApi.Services
             var restaurant = _dbContext.Restaurants
                 .FirstOrDefault(r => r.Id == id);
 
-            if (restaurant is null) throw new NotFoundException("Restaurant not found"); ;
+            if (restaurant is null) throw new NotFoundException("Restaurant not found");
+
+            var authorizationResult = _authorizationService.AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             _dbContext.Restaurants.Remove(restaurant);
             _dbContext.SaveChanges();
